backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy>=1.26,<2.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
scipy>=1.12.0
pytest>=8.0.0
gdown>=5.1.0
fashn-human-parser>=0.1.0
fal-client>=0.5.0
python-dotenv>=1.0.0
playwright>=1.40.0
pytest-playwright>=0.4.0
